# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/emotions": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "emo数据管理"
                ],
                "summary": "emo 更新emo数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dict"
                        }
                    }
                ],
                "responses": {}
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "emo数据管理"
                ],
                "summary": "emo 更新数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dict"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "emo数据管理"
                ],
                "summary": "emo 创建数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dict"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "emo数据管理"
                ],
                "summary": "emo 删除数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Dict"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "model.Dict": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "en": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DictItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "model.DictItem": {
            "type": "object",
            "properties": {
                "dictCode": {
                    "description": "外键，关联 Dict",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "labelEn": {
                    "description": "英文名称",
                    "type": "string"
                },
                "labelZh": {
                    "description": "中文名称",
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "description": "启用状态",
                    "type": "string"
                },
                "value": {
                    "description": "数据值，系统用的",
                    "type": "string"
                }
            }
        },
...
    }
}